third_party/WebKit/LayoutTests/virtual/gpu/fast/canvas/canvas-scale-drawImage-shadow-expected.txt

Ensure correct behavior of canvas with drawImage+shadow after scaling. A blue and red checkered pattern should be displayed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
FAIL d[3] should be 255. Was 203.
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
FAIL d[3] should be 255. Was 203.
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 76
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
FAIL d[3] is not around 76 (actual: 61)
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 114
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 114
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 114
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 114
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 34
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 34
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 34
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 34
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 57
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 57
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 57
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 57
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 17
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 17
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 17
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 17
PASS successfullyParsed is true

TEST COMPLETE
